pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-explainer"
version = "1.0.0"
description = "AI-powered CLI command explainer with risk assessment"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["cli", "command", "ai", "explainer", "security", "terminal"]
authors = [
    {name = "CLI Explainer Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

dependencies = [
    "pyperclip>=1.8.2",
    "keyboard>=0.13.5",
    "rich>=13.7.0",
    "litellm>=1.40.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/ThisIsxm/cli-command-explainer"
Documentation = "https://github.com/ThisIsxm/cli-command-explainer#readme"
Repository = "https://github.com/ThisIsxm/cli-command-explainer"
Changelog = "https://github.com/ThisIsxm/cli-command-explainer/blob/main/CHANGELOG.md"
Issues = "https://github.com/ThisIsxm/cli-command-explainer/issues"

[project.scripts]
cli-explainer = "src.main:main"

[tool.setuptools]
packages = ["src", "src.capturer", "src.parser", "src.risk", "src.explainer", "src.ui"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''
